requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
python-multipart
openai>=1.12.0
PyJWT>=2.8.0
httpx>=0.27.0
